feat: Move books around shelves
If user choses a different category for a book in the dropdown
menu, the book shelf is updated on the server as well as in the
state of the app so that state is reflected right away. Because the
state is also stored on the server, it works across page refreshes.