feat: Search
Users are now able to go to a search page from the main listing page
and search for new books. As user type in the query, search results
are instantly displayed on the page. Users are also able to categorize
books from search page. Going back from search page to main page keeps
the state up to date, as well as, the state is maintained between refreshes.