feature: sync state with searched books
The state of the application from the main page is not reflected when
a user searches for books and the books which are on a shelf are
returned. This PR adds the books state to search component and search
component syncs the state taking the state from main page as source of
truth.